Streamline repository agent profiles
