+++ .github/workflows/repository-agents.yml
@@ -26,7 +26,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: dev-tools/agents
+        working-directory: dev-tools/agents/runner
     steps:
       - name: Checkout
         uses: actions/checkout@v7
@@ -42,20 +42,20 @@ jobs:
       - name: Configure isolated uv cache
         run: echo "UV_CACHE_DIR=$RUNNER_TEMP/repository-agents-uv-cache" >> "$GITHUB_ENV"
 
-      - name: Validate standalone helper lock
-        run: uv lock --script helpers.py --check
+      - name: Validate profile resolver lock
+        run: uv lock --script profile_resolver.py --check
 
       - name: Run tests
         run: python -m unittest discover -s tests -v
 
       - name: Check Python
         run: |
-          uv run --with ruff==0.16.2 ruff check helpers.py tests/test_helpers.py
-          uv run --with ruff==0.16.2 ruff format --check helpers.py tests/test_helpers.py
-          python -m compileall -q helpers.py tests
+          uv run --with ruff==0.16.2 ruff check profile_resolver.py tests/test_profile_resolver.py
+          uv run --with ruff==0.16.2 ruff format --check profile_resolver.py tests/test_profile_resolver.py
+          python -m compileall -q profile_resolver.py tests
 
       - name: Check shell syntax
-        run: bash -n run.sh runtime/harnesses/pi/exec.sh
+        run: bash -n run.sh harnesses/pi/exec.sh
 
       - name: Build the staged Pi sandbox
         run: |
@@ -67,3 +67,16 @@ jobs:
           docker build \
             --tag repository-agent-pi:ci \
             "$RUNNER_TEMP/repository-agent-stage/sandbox"
+
+      - name: Check ephemeral gateway authentication
+        run: |
+          set +e
+          output="$(MODEL_ID_TOP=smoke-model ./run.sh \
+            --profile dev-note-reviewer \
+            --task editorial \
+            </dev/null 2>&1)"
+          status="$?"
+          set -e
+          printf '%s\n' "$output"
+          test "$status" -eq 2
+          grep -F "provider 'model-gateway' is not configured" <<<"$output"
