+++ dev-tools/agents/AGENTS.md
@@ -9,7 +9,7 @@
   a profile.
 - Disable automatic Pi resource and OpenShell provider discovery. Load only
   explicitly declared tools, skills, prompts, schemas, and providers.
-- Keep `helpers.py` a standalone utility; do not turn this tool into a Python
-  package without a concrete need for reusable Python APIs.
+- Keep `runner/profile_resolver.py` a standalone utility; do not turn this tool
+  into a Python package without a concrete need for reusable Python APIs.
 - Run the unittest, Ruff, compile, and shell syntax checks documented in
   `README.md` before handing off changes.
