+++ dev-tools/agents/profiles/dev-note-reviewer/profile.yaml
@@ -3,28 +3,24 @@ id: dev-note-reviewer
 display_name: Dev Note Reviewer
 description: Review OpenShell Dev Notes for editorial quality and technical grounding.
 
-harness:
-  name: pi
+harness: pi
 
 inference:
   provider: model-gateway
   context_window: 1000000
   max_tokens: 128000
 
-sandbox:
-  policy: policy.yaml
+policy: policy.yaml
 
 tasks:
   editorial:
     prompt: prompts/editorial.md
-    output_schema: schemas/review-response.json
+    output_schema: review-response.schema.json
+    output_schema_definition: editorial
     input_label: Untrusted projected prose data
-    tools: []
-    skills: []
 
   technical:
     prompt: prompts/technical.md
-    output_schema: schemas/review-response.json
+    output_schema: review-response.schema.json
+    output_schema_definition: technical
     input_label: Untrusted candidate data and diff
-    tools: []
-    skills: []
