+++ dev-tools/agents/runner/profile.schema.json
@@ -1,21 +1,16 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/NVIDIA/OpenShell-Research/dev-tools/agents/schemas/profile.schema.json",
+  "$id": "https://github.com/NVIDIA/OpenShell-Research/dev-tools/agents/runner/profile.schema.json",
   "title": "Repository agent profile",
   "type": "object",
   "additionalProperties": false,
-  "required": ["version", "id", "display_name", "description", "harness", "inference", "sandbox", "tasks"],
+  "required": ["version", "id", "display_name", "description", "harness", "inference", "policy", "tasks"],
   "properties": {
     "version": { "const": 1 },
     "id": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$" },
     "display_name": { "type": "string", "minLength": 1, "maxLength": 128 },
     "description": { "type": "string", "minLength": 1, "maxLength": 1024 },
-    "harness": {
-      "type": "object",
-      "additionalProperties": false,
-      "required": ["name"],
-      "properties": { "name": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$" } }
-    },
+    "harness": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$" },
     "inference": {
       "type": "object",
       "additionalProperties": false,
@@ -26,12 +21,7 @@
         "max_tokens": { "type": "integer", "minimum": 1 }
       }
     },
-    "sandbox": {
-      "type": "object",
-      "additionalProperties": false,
-      "required": ["policy"],
-      "properties": { "policy": { "type": "string", "minLength": 1 } }
-    },
+    "policy": { "type": "string", "minLength": 1 },
     "tasks": {
       "type": "object",
       "minProperties": 1,
@@ -42,6 +32,7 @@
         "properties": {
           "prompt": { "type": "string", "minLength": 1 },
           "output_schema": { "type": "string", "minLength": 1 },
+          "output_schema_definition": { "type": "string", "pattern": "^[A-Za-z][A-Za-z0-9]*$" },
           "input_label": { "type": "string", "minLength": 1, "maxLength": 200 },
           "tools": { "type": "array", "uniqueItems": true, "items": { "type": "string", "pattern": "^[a-z][a-z0-9_-]*$" } },
           "skills": { "type": "array", "uniqueItems": true, "items": { "type": "string", "minLength": 1 } },
