+++ dev-tools/agents/runner/provider.schema.json
@@ -1,0 +1,37 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "https://github.com/NVIDIA/OpenShell-Research/dev-tools/agents/runner/provider.schema.json",
+  "title": "Repository agent inference provider",
+  "type": "object",
+  "additionalProperties": false,
+  "required": ["version", "id", "description", "openshell", "sandbox"],
+  "properties": {
+    "version": { "const": 1 },
+    "id": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$" },
+    "description": { "type": "string", "minLength": 1, "maxLength": 1024 },
+    "openshell": {
+      "type": "object",
+      "additionalProperties": false,
+      "required": ["type", "base_url_source_env", "api_key_source_env", "base_url_config_key"],
+      "properties": {
+        "type": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$" },
+        "base_url_source_env": { "$ref": "#/$defs/environmentName" },
+        "api_key_source_env": { "$ref": "#/$defs/environmentName" },
+        "base_url_config_key": { "$ref": "#/$defs/environmentName" }
+      }
+    },
+    "sandbox": {
+      "type": "object",
+      "additionalProperties": false,
+      "required": ["api", "base_url", "api_key"],
+      "properties": {
+        "api": { "type": "string", "minLength": 1 },
+        "base_url": { "type": "string", "minLength": 1 },
+        "api_key": { "type": "string", "minLength": 1 }
+      }
+    }
+  },
+  "$defs": {
+    "environmentName": { "type": "string", "pattern": "^[A-Z][A-Z0-9_]*$" }
+  }
+}
